feat: добавляет оптимизацию для главной стр
